tidy timetable code, make the page reactive and move it to pages not component
